Updated file organization for `ast` packages
