--- language/ast/selection.go
@@ -1,9 +1,0 @@
-package ast
-
-type Selection interface {
-}
-
-// Ensure that all definition types implements Selection interface
-var _ Selection = (*Field)(nil)
-var _ Selection = (*FragmentSpread)(nil)
-var _ Selection = (*InlineFragment)(nil)
